requirements.txt
openai
python-dotenv
openpyxl
streamlit